test: use a module instead of a bundle to test multiple versions (#193)
Since Jahia/jahia-private#3626, only modules and fragments can be installed.
Meaning the test to check that multiple versions of the same bundle are reported as duplicates, is not relevant anymore (and can't be tested as the installation of a bundle is no longer possible in the setup of the test).